% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnik.R
\docType{package}
\name{birdnik}
\alias{birdnik}
\alias{birdnik-package}
\title{Connector to the Wordnik API}
\description{
This package provides functions for the Wordnik API.
}
\seealso{
\code{\link{word_definitions}}, \code{\link{word_top_example}},
\code{\link{related_words}}, \code{\link{word_pronunciations}},
\code{\link{word_frequency}}, \code{\link{word_bigrams}}, and
\code{\link{birdnik_get}}.
}

